authentication for posts

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,10 @@ use App\Post;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -40,10 +44,12 @@ class PostsController extends Controller
             'myTitle' => 'required',
             'body' => 'required'
         ]);
+        $id = auth()->user()->id;
 
         $post = new Post;
         $post->title = $request->input('myTitle');
         $post->body = $request->input('body');
+        $post->user_id = $id;
         $post->save();
 
         return redirect( route('posts.index') );
@@ -69,7 +75,11 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect( route('posts.index') );
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -106,6 +116,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect( route('posts.index') );
+        }
         $post->delete();
         return redirect( route('posts.index'));
     }
